+++ src/wrapper.hpp
@@ -21,23 +21,23 @@ public:
         virtual ~FunctionalWrapper()
         { }
 
-        virtual double operator()(const double* data, const size_t length) const = 0;
+        virtual float operator()(const float* data, const size_t length) const = 0;
 };
 
 // Simple wrapper without any additional arguments
 class SimpleFunctionalWrapper : public FunctionalWrapper {
 public:
-        SimpleFunctionalWrapper(std::function<double(const double*, const size_t)> function)
+        SimpleFunctionalWrapper(std::function<float(const float*, const size_t)> function)
                 : _function(function)
         { }
 
-        double operator()(const double* data, const size_t length) const
+        float operator()(const float* data, const size_t length) const
         {
                 return _function(data, length);
         }
 
 private:
-        const std::function<double(const double*, const size_t)> _function;
+        const std::function<float(const float*, const size_t)> _function;
 };
 
 // Generic wrapper implementation using variadic templates
@@ -45,7 +45,7 @@ template<typename... Parameters>
 class GenericFunctionalWrapper : public FunctionalWrapper
 {
 public:
-        GenericFunctionalWrapper(std::function<double(const double*, const size_t, Parameters...)> functional)
+        GenericFunctionalWrapper(std::function<float(const float*, const size_t, Parameters...)> functional)
                 : _functional(functional)
         {
                 // FIXME: it is possible to pass an invalid functional not matching Parameters...
@@ -57,15 +57,15 @@ public:
                 _configured_functional = std::bind(_functional, std::placeholders::_1, std::placeholders::_2, parameters...);
         }
 
-        double operator()(const double* data, const size_t length) const
+        float operator()(const float* data, const size_t length) const
         {
                 assert(_configured_functional);
                 return _configured_functional(data, length);
         }
 
 private:
-        const std::function<double(const double*, const size_t, Parameters...)> _functional;
-        std::function<double(const double*, const size_t)> _configured_functional;
+        const std::function<float(const float*, const size_t, Parameters...)> _functional;
+        std::function<float(const float*, const size_t)> _configured_functional;
 };
 
 #endif
